@@ -4,10 +4,17 @@ using Samples.Specifications.Server.Domain.Entities;
 
 namespace Samples.Specifications.Server.Api.Mappers
 {
-    internal static class UserMapper
+    public class UserMapper
     {
-        internal static User MapToUser(UserDto userDto) => Mapper.Map<User>(userDto);
+        private readonly IMapper _mapper;
 
-        internal static UserDto MapToUserDto(User user) => Mapper.Map<UserDto>(user);
+        public UserMapper(IMapper mapper)
+        {
+            _mapper = mapper;
+        }
+
+        internal User MapToUser(UserDto userDto) => _mapper.Map<User>(userDto);
+
+        internal UserDto MapToUserDto(User user) => _mapper.Map<UserDto>(user);
     }
 }
