@@ -4,12 +4,19 @@ using Samples.Specifications.Server.Domain.Entities;
 
 namespace Samples.Specifications.Server.Api.Mappers
 {
-    internal static class WarehouseMapper
+    public class WarehouseMapper
     {
-        internal static WarehouseItem MapToWarehouseItem(WarehouseItemDto warehouseItemDto) =>
-            Mapper.Map<WarehouseItem>(warehouseItemDto);
+        private readonly IMapper _mapper;
 
-        internal static WarehouseItemDto MapToWarehouseItemDto(WarehouseItem warehouseItem) =>
-            Mapper.Map<WarehouseItemDto>(warehouseItem);
+        public WarehouseMapper(IMapper mapper)
+        {
+            _mapper = mapper;
+        }
+
+        internal WarehouseItem MapToWarehouseItem(WarehouseItemDto warehouseItemDto) =>
+            _mapper.Map<WarehouseItem>(warehouseItemDto);
+
+        internal WarehouseItemDto MapToWarehouseItemDto(WarehouseItem warehouseItem) =>
+            _mapper.Map<WarehouseItemDto>(warehouseItem);
     }
 }
