@@ -3,8 +3,8 @@ using System.Reflection;
 using JetBrains.Annotations;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.PlatformAbstractions;
+using Microsoft.OpenApi.Models;
 using Solid.Practices.Modularity;
-using Swashbuckle.AspNetCore.Swagger;
 
 namespace Samples.Specifications.Server.Api
 {
@@ -13,7 +13,7 @@ namespace Samples.Specifications.Server.Api
     {
         public void RegisterModule(IServiceCollection dependencyRegistrator) => dependencyRegistrator.AddSwaggerGen(c =>
         {
-            c.SwaggerDoc("v1", new Info {Title = "My API", Version = "v1"});
+            c.SwaggerDoc("v1", new OpenApiInfo {Title = "My API", Version = "v1"});
             c.IncludeXmlComments(Path.Combine(PlatformServices.Default.Application.ApplicationBasePath,
                 $"{Assembly.GetExecutingAssembly().GetName().Name}.xml"));
         });
